bpf_loader: extract cache deploy to program-runtime
