+++ program-runtime/src/lib.rs
@@ -13,6 +13,7 @@
 
 pub use solana_sbpf;
 pub mod cpi;
+pub mod deploy;
 pub mod execution_budget;
 pub mod invoke_context;
 pub mod loaded_programs;
