+++ programs/bpf_loader/src/lib.rs
@@ -15,20 +15,16 @@ use qualifier_attr::qualifiers;
 use {
     cfg_if::cfg_if,
     solana_bincode::limited_deserialize,
-    solana_clock::Slot,
     solana_instruction::{AccountMeta, error::InstructionError},
     solana_loader_v3_interface::{
         instruction::UpgradeableLoaderInstruction, state::UpgradeableLoaderState,
     },
     solana_program_entrypoint::{MAX_PERMITTED_DATA_INCREASE, SUCCESS},
     solana_program_runtime::{
+        deploy_program,
         execution_budget::MAX_INSTRUCTION_STACK_DEPTH,
         invoke_context::{BpfAllocator, InvokeContext, SerializedAccountMetadata, SyscallContext},
-        loaded_programs::{
-            DELAY_VISIBILITY_SLOT_OFFSET, LoadProgramMetrics, ProgramCacheEntry,
-            ProgramCacheEntryOwner, ProgramCacheEntryType, ProgramCacheForTxBatch,
-            ProgramRuntimeEnvironment,
-        },
+        loaded_programs::{ProgramCacheEntry, ProgramCacheEntryOwner, ProgramCacheEntryType},
         mem_pool::VmMemoryPool,
         serialization, stable_log,
         sysvar_cache::get_sysvar_with_account_check,
@@ -37,11 +33,9 @@ use {
     solana_sbpf::{
         declare_builtin_function,
         ebpf::{self, MM_HEAP_START},
-        elf::{ElfError, Executable},
+        elf::Executable,
         error::{EbpfError, ProgramResult},
         memory_region::{AccessType, MemoryMapping, MemoryRegion},
-        program::BuiltinProgram,
-        verifier::RequisiteVerifier,
         vm::{ContextObject, EbpfVm},
     },
     solana_sdk_ids::{
@@ -49,7 +43,7 @@ use {
     },
     solana_svm_log_collector::{LogCollector, ic_logger_msg, ic_msg},
     solana_svm_measure::measure::Measure,
-    solana_svm_type_overrides::sync::{Arc, atomic::Ordering},
+    solana_svm_type_overrides::sync::Arc,
     solana_system_interface::{MAX_PERMITTED_DATA_LENGTH, instruction as system_instruction},
     solana_transaction_context::{
         IndexOfAccount, TransactionContext, instruction::InstructionContext,
@@ -66,158 +60,6 @@ const UPGRADEABLE_LOADER_COMPUTE_UNITS: u64 = 2_370;
 
 thread_local! {
     pub static MEMORY_POOL: RefCell<VmMemoryPool> = RefCell::new(VmMemoryPool::new());
-}
-
-fn morph_into_deployment_environment_v1<'a>(
-    from: Arc<BuiltinProgram<InvokeContext<'a, 'a>>>,
-) -> Result<BuiltinProgram<InvokeContext<'a, 'a>>, ElfError> {
-    let mut config = from.get_config().clone();
-    config.reject_broken_elfs = true;
-    // Once the tests are being build using a toolchain which supports the newer SBPF versions,
-    // the deployment of older versions will be disabled:
-    // config.enabled_sbpf_versions =
-    //     *config.enabled_sbpf_versions.end()..=*config.enabled_sbpf_versions.end();
-
-    let mut result = BuiltinProgram::new_loader(config);
-
-    for (_key, (name, value)) in from.get_function_registry().iter() {
-        // Deployment of programs with sol_alloc_free is disabled. So do not register the syscall.
-        if name != *b"sol_alloc_free_" {
-            result.register_function(unsafe { std::str::from_utf8_unchecked(name) }, value)?;
-        }
-    }
-
-    Ok(result)
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn load_program_from_bytes(
-    log_collector: Option<Rc<RefCell<LogCollector>>>,
-    load_program_metrics: &mut LoadProgramMetrics,
-    programdata: &[u8],
-    loader_key: &Pubkey,
-    account_size: usize,
-    deployment_slot: Slot,
-    program_runtime_environment: Arc<BuiltinProgram<InvokeContext<'static, 'static>>>,
-    reloading: bool,
-) -> Result<ProgramCacheEntry, InstructionError> {
-    let effective_slot = deployment_slot.saturating_add(DELAY_VISIBILITY_SLOT_OFFSET);
-    let loaded_program = if reloading {
-        // Safety: this is safe because the program is being reloaded in the cache.
-        unsafe {
-            ProgramCacheEntry::reload(
-                loader_key,
-                program_runtime_environment,
-                deployment_slot,
-                effective_slot,
-                programdata,
-                account_size,
-                load_program_metrics,
-            )
-        }
-    } else {
-        ProgramCacheEntry::new(
-            loader_key,
-            program_runtime_environment,
-            deployment_slot,
-            effective_slot,
-            programdata,
-            account_size,
-            load_program_metrics,
-        )
-    }
-    .map_err(|err| {
-        ic_logger_msg!(log_collector, "{}", err);
-        InstructionError::InvalidAccountData
-    })?;
-    Ok(loaded_program)
-}
-
-/// Directly deploy a program using a provided invoke context.
-/// This function should only be invoked from the runtime, since it does not
-/// provide any account loads or checks.
-pub fn deploy_program(
-    log_collector: Option<Rc<RefCell<LogCollector>>>,
-    program_cache_for_tx_batch: &mut ProgramCacheForTxBatch,
-    program_runtime_environment: ProgramRuntimeEnvironment,
-    program_id: &Pubkey,
-    loader_key: &Pubkey,
-    account_size: usize,
-    programdata: &[u8],
-    deployment_slot: Slot,
-) -> Result<LoadProgramMetrics, InstructionError> {
-    let mut load_program_metrics = LoadProgramMetrics::default();
-    let mut register_syscalls_time = Measure::start("register_syscalls_time");
-    let deployment_program_runtime_environment =
-        morph_into_deployment_environment_v1(program_runtime_environment.clone()).map_err(|e| {
-            ic_logger_msg!(log_collector, "Failed to register syscalls: {}", e);
-            InstructionError::ProgramEnvironmentSetupFailure
-        })?;
-    register_syscalls_time.stop();
-    load_program_metrics.register_syscalls_us = register_syscalls_time.as_us();
-    // Verify using stricter deployment_program_runtime_environment
-    let mut load_elf_time = Measure::start("load_elf_time");
-    let executable = Executable::<InvokeContext>::load(
-        programdata,
-        Arc::new(deployment_program_runtime_environment),
-    )
-    .map_err(|err| {
-        ic_logger_msg!(log_collector, "{}", err);
-        InstructionError::InvalidAccountData
-    })?;
-    load_elf_time.stop();
-    load_program_metrics.load_elf_us = load_elf_time.as_us();
-    let mut verify_code_time = Measure::start("verify_code_time");
-    executable.verify::<RequisiteVerifier>().map_err(|err| {
-        ic_logger_msg!(log_collector, "{}", err);
-        InstructionError::InvalidAccountData
-    })?;
-    verify_code_time.stop();
-    load_program_metrics.verify_code_us = verify_code_time.as_us();
-    // Reload but with program_runtime_environment
-    let executor = load_program_from_bytes(
-        log_collector,
-        &mut load_program_metrics,
-        programdata,
-        loader_key,
-        account_size,
-        deployment_slot,
-        program_runtime_environment,
-        true,
-    )?;
-    if let Some(old_entry) = program_cache_for_tx_batch.find(program_id) {
-        executor.tx_usage_counter.store(
-            old_entry.tx_usage_counter.load(Ordering::Relaxed),
-            Ordering::Relaxed,
-        );
-    }
-    load_program_metrics.program_id = program_id.to_string();
-    program_cache_for_tx_batch.store_modified_entry(*program_id, Arc::new(executor));
-    Ok(load_program_metrics)
-}
-
-#[macro_export]
-macro_rules! deploy_program {
-    ($invoke_context:expr, $program_id:expr, $loader_key:expr, $account_size:expr, $programdata:expr, $deployment_slot:expr $(,)?) => {
-        assert_eq!(
-            $deployment_slot,
-            $invoke_context.program_cache_for_tx_batch.slot()
-        );
-        let load_program_metrics = $crate::deploy_program(
-            $invoke_context.get_log_collector(),
-            $invoke_context.program_cache_for_tx_batch,
-            $invoke_context
-                .get_program_runtime_environments_for_deployment()
-                .program_runtime_v1
-                .clone(),
-            $program_id,
-            $loader_key,
-            $account_size,
-            $programdata,
-            $deployment_slot,
-        )?;
-        load_program_metrics.submit_datapoint(&mut $invoke_context.timings);
-    };
 }
 
 fn write_program_data(
@@ -1708,8 +1550,13 @@ fn execute<'a, 'b: 'a>(
 mod test_utils {
     #[cfg(feature = "svm-internal")]
     use {
-        super::*, agave_syscalls::create_program_runtime_environment_v1,
-        solana_account::ReadableAccount, solana_loader_v4_interface::state::LoaderV4State,
+        super::*,
+        agave_syscalls::create_program_runtime_environment_v1,
+        solana_account::ReadableAccount,
+        solana_loader_v4_interface::state::LoaderV4State,
+        solana_program_runtime::{
+            deploy::load_program_from_bytes, loaded_programs::LoadProgramMetrics,
+        },
         solana_sdk_ids::loader_v4,
     };
 
@@ -1792,7 +1639,9 @@ mod tests {
         },
         solana_pubkey::Pubkey,
         solana_rent::Rent,
+        solana_sbpf::program::BuiltinProgram,
         solana_sdk_ids::{system_program, sysvar},
+        solana_svm_type_overrides::sync::atomic::Ordering,
         std::{fs::File, io::Read, ops::Range, sync::atomic::AtomicU64},
     };
