+++ programs/loader-v4/Cargo.toml
@@ -16,7 +16,9 @@ crate-type = ["lib"]
 name = "solana_loader_v4_program"
 
 [features]
+default = ["metrics"]
 agave-unstable-api = []
+metrics = ["solana-program-runtime/metrics"]
 shuttle-test = [
     "solana-program-runtime/shuttle-test",
     "solana-sbpf/shuttle-test",
